fastAPI/requirements.txt
fastapi
uvicorn
python-jose
passlib
python-multipart
sqlalchemy
psycopg2-binary
python-dotenv
google-generativeai
pydantic
pydantic-settings
python-dateutil
requests
aiofiles
httpx
openai
pydantic[email]
bcrypt